pyproject.toml
[project]
name = "autotrust"
version = "0.1.0"
description = "AutoEmailTrust v3.5: Automated email trust scoring research loop"
requires-python = ">=3.12"
dependencies = [
    "anthropic",
    "openai",
    "ollama",
    "python-dotenv",
    "pydantic",
    "pyyaml",
    "gitpython",
    "httpx",
    "rich",
    "structlog",
    "datasets",
    "scikit-learn",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["autotrust"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100
